Move config layering.
